Move repair_response to core
