+++ core/src/lib.rs
@@ -36,6 +36,7 @@ pub mod poh_recorder;
 pub mod poh_service;
 pub mod progress_map;
 pub mod pubkey_references;
+pub mod repair_response;
 pub mod repair_service;
 pub mod replay_stage;
 pub mod request_response;
